Optimization

+++ client/src/app/messages/messages.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { MessageService } from '../_services/message.service';
 import { IPagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 
 @Component({
 	selector: 'app-messages',
@@ -16,7 +17,10 @@ export class MessagesComponent implements OnInit {
 	pageSize = 5;
 	loading = false;
 
-	constructor(private messageService: MessageService) {}
+	constructor(
+		private messageService: MessageService,
+		private confirmService: ConfirmService
+	) {}
 
 	ngOnInit(): void {
 		this.loadMessages();
@@ -38,11 +42,17 @@ export class MessagesComponent implements OnInit {
 	}
 
 	deleteMessage(id: number) {
-		this.messageService.deleteMessage(id).subscribe(() => {
-			this.messages.splice(
-				this.messages.findIndex((x) => x.id === id),
-				1
-			);
-		});
+		this.confirmService
+			.confirm('Confirm delete message')
+			.subscribe((result) => {
+				if (result) {
+					this.messageService.deleteMessage(id).subscribe(() => {
+						this.messages.splice(
+							this.messages.findIndex((x) => x.id === id),
+							1
+						);
+					});
+				}
+			});
 	}
 }
